feat: Add new routes for different components
- Add new routes for Home, 3d, Device List, Edit Device, Add Device, Models List, Edit Model, Add Model, Attribute Dictionary List, Add Attribute Dictionary, Edit Attribute Dictionary, Attribute List, Add Attribute, Edit Attribute, Connections List, Add Connections, and Edit Connections components.

The commit also includes changes to the `app-routing.module.ts` file to import `Title` from `@angular/platform-browser` and inject it into the `TemplatePageTitleStrategy` class to set the page title dynamically based on the current route.

The commit also includes changes to the `attribute-list.component.ts` file to define an `@Input` property `attributeComponent`, and rename the `LoadAttribute` function to `LoadAttributes`.

The commit also includes changes to the `edit-device.component.ts` file to initialize the `attributeComponent` property to an empty string.

The commit also includes changes to the `log.component.ts` file to remove unnecessary imports and commented out code, and update the `loadLog` and `findNameInLogMessage` functions.

The commit also includes changes to the `attribute.service.ts` file to update the return type of the `GetAttributes` function to `Observable<Attribute[]>`.