feat: Add home route to app-routing
Add a new route to the `app-routing.module.ts` file for the home page. This route is assigned to the `HomeComponent` and has the title "Home".

mod: Update container class to container-fluid

Update the `container` class to `container-fluid` in the `app.component.html` file to use the fluid container layout.

mod: Update JSON.stringify with indentation

Update the `toString` method in the `add-attribute-dictionary.component.ts` and `edit-attribute-dictionary.component.ts` files to use JSON.stringify with an indentation of 2 spaces.

mod: Update buttons layout in attribute-dictionary-list

Update the layout of the buttons in the `attribute-dictionary-list.component.html` file to align them to the right.

mod: Update container class to container-fluid in UI components

Update the `container` class to `container-fluid` in the `add-attribute.component.html`, `attribute-list.component.html`, and `edit-attribute.component.html` files to use the fluid container layout.